More accuracy.

+++ C#/Benchmark.cs
@@ -95,7 +95,7 @@ public static class Benchmark
         if (managedElapsedMillisecond.HasValue && nativeElapsedMillisecond.HasValue)
         {
             Console.ForegroundColor = ConsoleColor.Green;
-            Console.WriteLine("\nRatio: {0:0.4}", managedElapsedMillisecond / nativeElapsedMillisecond);
+            Console.WriteLine("\nRatio: {0:0.4}", (double)managedElapsedMillisecond / nativeElapsedMillisecond);
             Console.ForegroundColor = ConsoleColor.Gray;
         }
